Update Part 2.9: The Phonebook Step4
I spent too much time with this step trying to manage the persons' list data with the state. Until I realized that I needed to define a source in where I should manage the data because the purpose of the React's state isn't be a database.
I'm proposing this extra information for this step because I think it makes it clearer how to approach the challenge, as well as can be a good way to bridge to Part 2c.

P.D.: I'm really enjoying the course. Thank you very much!